cannasoltech_automation/functions/requirements.txt
functions-framework
firebase-functions
firebase-admin
openpyxl
google-cloud-storage
google-genai
google-auth 
google-auth-oauthlib